scripts/phase1/requirements.txt
pyodbc==5.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.4
